+++ src/components/screens/Post/components/PostTitle.tsx
@@ -1,0 +1,23 @@
+import { Text, useTheme } from "native-base";
+import React from "react";
+import { useAppSelector } from "../../../../../store";
+import { selectSettings } from "../../../../slices/settings/settingsSlice";
+
+function Title({ title, mt, mb }: { title: string; mt: number; mb: number }) {
+  const theme = useTheme();
+  const { fontWeightPostTitle } = useAppSelector(selectSettings);
+  return (
+    <Text
+      mt={mt}
+      mb={mb}
+      mx={4}
+      fontSize="lg"
+      fontWeight={fontWeightPostTitle}
+      color={theme.colors.app.textPrimary}
+    >
+      {title}
+    </Text>
+  );
+}
+
+export default Title;
