+++ src/hooks/post/postHooks.ts
@@ -33,6 +33,9 @@ export interface UsePost {
   sortType: CommentSortType;
   setSortType: React.Dispatch<SetStateAction<CommentSortType>>;
 
+  collapsed: boolean;
+  setCollapsed: React.Dispatch<SetStateAction<boolean>>;
+
   showLoadAll: boolean;
   setShowLoadAll: React.Dispatch<SetStateAction<boolean>>;
 
@@ -55,6 +58,8 @@ const usePost = (commentId: string | null): UsePost => {
   const [commentsLoading, setCommentsLoading] = useState<boolean>(true);
   const [commentsError, setCommentsError] = useState<boolean>(false);
   const [currentPost, setCurrentPost] = useState<PostView>(post);
+
+  const [collapsed, setCollapsed] = useState<boolean>(false);
 
   const [sortType, setSortType] = useState<CommentSortType>("Top");
   const [showLoadAll, setShowLoadAll] = useState(true);
@@ -279,6 +284,9 @@ const usePost = (commentId: string | null): UsePost => {
 
     showLoadAll,
     setShowLoadAll,
+
+    collapsed,
+    setCollapsed,
 
     doSave,
